Rename StandardSQLErrorCode
