+++ shardingsphere-proxy/shardingsphere-proxy-frontend/shardingsphere-proxy-frontend-mysql/src/test/java/org/apache/shardingsphere/proxy/frontend/mysql/err/MySQLErrPacketFactoryTest.java
@@ -27,7 +27,7 @@ import org.apache.shardingsphere.infra.exception.dialect.DBCreateExistsException
 import org.apache.shardingsphere.infra.exception.dialect.DBDropNotExistsException;
 import org.apache.shardingsphere.infra.exception.dialect.TableModifyInTransactionException;
 import org.apache.shardingsphere.proxy.backend.handler.distsql.ral.common.exception.UnsupportedVariableException;
-import org.apache.shardingsphere.proxy.frontend.exception.UnsupportedCommandException;
+import org.apache.shardingsphere.infra.exception.UnsupportedCommandException;
 import org.apache.shardingsphere.proxy.frontend.exception.UnsupportedPreparedStatementException;
 import org.apache.shardingsphere.infra.exception.dialect.NoSuchTableException;
 import org.apache.shardingsphere.infra.exception.dialect.TableExistsException;
@@ -160,7 +160,7 @@ public final class MySQLErrPacketFactoryTest {
         assertThat(actual.getSequenceId(), is(1));
         assertThat(actual.getErrorCode(), is(1235));
         assertThat(actual.getSqlState(), is("42000"));
-        assertThat(actual.getErrorMessage(), is("This version of ShardingSphere-Proxy doesn't yet support this SQL. 'No reason'"));
+        assertThat(actual.getErrorMessage(), is("Unsupported SQL: No reason"));
     }
     
     @Test
@@ -178,7 +178,7 @@ public final class MySQLErrPacketFactoryTest {
         assertThat(actual.getSequenceId(), is(1));
         assertThat(actual.getErrorCode(), is(1000));
         assertThat(actual.getSqlState(), is("C1000"));
-        assertThat(actual.getErrorMessage(), is("Circuit break mode is ON."));
+        assertThat(actual.getErrorMessage(), is("Circuit break mode is ON"));
     }
     
     @Test
@@ -200,7 +200,7 @@ public final class MySQLErrPacketFactoryTest {
         assertThat(actual.getSequenceId(), is(1));
         assertThat(actual.getErrorCode(), is(1998));
         assertThat(actual.getSqlState(), is("C1998"));
-        assertThat(actual.getErrorMessage(), is("Unsupported command: [No reason]"));
+        assertThat(actual.getErrorMessage(), is("Unsupported command: No reason"));
     }
     
     @Test
@@ -209,15 +209,6 @@ public final class MySQLErrPacketFactoryTest {
         assertThat(actual.getSequenceId(), is(1));
         assertThat(actual.getErrorCode(), is(1999));
         assertThat(actual.getSqlState(), is("C1999"));
-        assertThat(actual.getErrorMessage(), is("Unknown exception: [No reason]"));
-    }
-    
-    @Test
-    public void assertNewInstanceWithRuntimeException() {
-        MySQLErrPacket actual = MySQLErrPacketFactory.newInstance(new RuntimeException("No reason"));
-        assertThat(actual.getSequenceId(), is(1));
-        assertThat(actual.getErrorCode(), is(1997));
-        assertThat(actual.getSqlState(), is("C1997"));
-        assertThat(actual.getErrorMessage(), is("Runtime exception: [No reason]"));
+        assertThat(actual.getErrorMessage(), is("Unknown exception: No reason"));
     }
 }
